* develop:
  remove old icon
  new icon
  Implemented: - new app icon - removed hard coded numbers and implemented enum - fixed bugs with segmented search radius
  - Implemented web view - Passed detail object to show first weblink of detail object - Sorted out segue loading weather and detail

# Conflicts:
#	DiveAdvisor/DiveAdvisor.xcodeproj/project.pbxproj

+++ DiveAdvisor/DAServiceClass.swift
@@ -33,7 +33,7 @@ class DAServiceClass {
         }
     }
     
-    static func diveSearchByGeo(lat: Double, lng: Double, dist: Int)  {
+    static func diveSearchByGeo(lat: Double, lng: Double, dist: Double)  {
         let url = DAUrlCreator.createDAURLWithComponents(term: .bySearchCoordDist(lat, lng, dist))
         alamoFireCall(url: url!, matchOn: .sites) {
             (result, sites, aNSerror) -> Void in
